move Report view & filter to 'reports' app
