+++ analysts/urls.py
@@ -1,6 +1,8 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
+
 from .views import *
+from reports.views import AnalystReportsView
 
 router = DefaultRouter()
